requirements.txt
altgraph==0.17.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
colour==0.1.5
macholib==1.16
numpy==1.23.1
opencv-python==4.6.0.66
Pillow==9.2.0
pyinstaller @ https://github.com/pyinstaller/pyinstaller/archive/develop.zip
pyinstaller-hooks-contrib==2022.8
pyserial==3.5
tkmacosx==1.0.5
